+++ src/app/zone/zone.component.ts
@@ -4,6 +4,7 @@ import {MatSort} from '@angular/material/sort';
 import {MatTableDataSource} from '@angular/material/table';
 import { BackendconnectionService } from '../backendconnection.service';
 import {Router} from '@angular/router';
+import Swal from 'sweetalert2/dist/sweetalert2.js'; 
 
 
 export interface UserData {
@@ -123,16 +124,45 @@ export class ZoneComponent implements AfterViewInit {
 
   zoneDelete(x){
 
-     alert(x)
-    if (confirm('Are you sure to delete this record ?') == true) {
+    //  alert(x)
+    // if (confirm('Are you sure to delete this record ?') == true) {
+      
 
-    this.backend.deletezone(x)
+   
+      Swal.fire({
+        title: 'Are you sure?',
+        text: 'You will not be able to recover this Zone!',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'Yes, delete it!',
+        cancelButtonText: 'No, keep it'
+      }).then((result) => {
+        if (result.value) {
+
+           this.backend.deletezone(x)
     .subscribe((data)=> { 
 
       //  console.log(data);
       this.getAllZones();
+      Swal.fire(
+        'Deleted!',
+        'Zone has been deleted.',
+        'success'
+      )
     });
-  }
+          
+        // For more information about handling dismissals please visit
+        // https://sweetalert2.github.io/#handling-dismissals
+        } else if (result.dismiss === Swal.DismissReason.cancel) {
+          Swal.fire(
+            'Cancelled',
+            'Zone is safe :)',
+            'error'
+          )
+        }
+      }) 
+   
+  // }
   }
